update(auth-context) removed state provider from login screen
