+++ context/AuthContextProvider.js
@@ -3,6 +3,8 @@ import {
   signInWithEmailAndPassword,
   createUserWithEmailAndPassword,
 } from "firebase/auth";
+
+import { onAuthStateChanged } from "firebase/auth";
 import { auth, db } from "../services/firebaseConfig";
 import {
   doc,
@@ -13,7 +15,7 @@ import {
   getDocs,
 } from "firebase/firestore";
 
-import RequestBuilder from "../builders/RequestBuilder"
+import RequestBuilder from "../builders/RequestBuilder";
 import createStudent from "../builders/StudentBuilder";
 
 const AuthContext = createContext();
@@ -23,7 +25,6 @@ const AuthContext = createContext();
 // }
 
 function AuthProvider({ children }) {
-
   const [activeUser, setActiveUser] = useState("");
   const [loading, setLoading] = useState(true);
   const [isTutor, setIsTutor] = useState(false);
@@ -48,43 +49,50 @@ function AuthProvider({ children }) {
   function signOut() {
     return auth.signOut();
   }
- // create a new student
-const signUp = async (email, password, userName) => {
-  try {
-    const userCredential = await createUserWithEmailAndPassword(auth, email, password);
-    const user = userCredential.user;
-    const userId = user.uid;
+  // create a new student
+  const signUp = async (email, password, userName) => {
+    try {
+      const userCredential = await createUserWithEmailAndPassword(
+        auth,
+        email,
+        password
+      );
+      const user = userCredential.user;
+      const userId = user.uid;
 
-    console.log('This is start of the function before execution')
-    // Create a new Student object using the StudentBuilder
-    const studentToStudentCollection = createStudent()
-    .withUid(userId)
-    .withEmail(email)
-    .withDisplayName(userName)
-    .build();
-    
-    const studentToUserCollection = createStudent()
-    .withEmail(email)
-    .withDisplayName(userName)
-    .build();
+      console.log("This is start of the function before execution");
+      // Create a new Student object using the StudentBuilder
+      const studentToStudentCollection = createStudent()
+        .withUid(userId)
+        .withEmail(email)
+        .withDisplayName(userName)
+        .build();
 
-    console.log('This is the student object: ',studentToStudentCollection, studentToUserCollection)
+      const studentToUserCollection = createStudent()
+        .withEmail(email)
+        .withDisplayName(userName)
+        .build();
 
-    // Add the student student to the student collection
-    const studentToStudentRef = doc(db, "students", userId);
-    await setDoc(studentToStudentRef, studentToStudentCollection);
+      console.log(
+        "This is the student object: ",
+        studentToStudentCollection,
+        studentToUserCollection
+      );
 
-    // add the student to the user collection
-    const studentTutorRef = doc(db, "users", userId);
-    await setDoc(studentTutorRef, studentToUserCollection);
+      // Add the student student to the student collection
+      const studentToStudentRef = doc(db, "students", userId);
+      await setDoc(studentToStudentRef, studentToStudentCollection);
 
-    return userCredential;
-  } catch (error) {
-    console.error("Registration error:", error.message);
-    // Handle the error as needed
-  }
-}; 
+      // add the student to the user collection
+      const studentTutorRef = doc(db, "users", userId);
+      await setDoc(studentTutorRef, studentToUserCollection);
 
+      return userCredential;
+    } catch (error) {
+      console.error("Registration error:", error.message);
+      // Handle the error as needed
+    }
+  };
 
   // Get the tutor status
   const getUserRole = async (currentUserId) => {
@@ -94,7 +102,7 @@ const signUp = async (email, password, userName) => {
       const userDoc = await getDoc(userRef);
 
       // Check if the user document exists and has the isTutor field set to true
-      if (userDoc.exists && userDoc.data().role === 'tutor') {
+      if (userDoc.exists && userDoc.data().role === "tutor") {
         setIsTutor(true);
         return true;
       } else {
@@ -117,9 +125,9 @@ const signUp = async (email, password, userName) => {
       const userDoc = await getDoc(userRef);
 
       // Check if the user document exists and has the display_name field
-      if (userDoc.exists && userDoc.data().display_name) {
-        const name = userDoc.data().display_name;
-        return name
+      if (userDoc.exists && userDoc.data().displayName) {
+        const name = userDoc.data().displayName;
+        return name;
       } else {
         // Return a default value or throw an error
         return "Unknown user";
@@ -135,11 +143,19 @@ const signUp = async (email, password, userName) => {
   };
 
   useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      setLoading(false);
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      if (user) {
+        console.log("User is signed in:", user.uid);
+        // Redirect or navigate to the home screen
+      } else {
+        console.log("User is signed out");
+      }
     });
 
-    return unsubscribe;
+    // Clean up the listener when the component unmounts
+    return () => {
+      unsubscribe();
+    };
   }, []);
 
   // const value = {
